fixes for api update and handle slave conditions
